requirements.txt
beautifulsoup4==4.12.0
eth_utils==1.10.0
neo4j==5.7.0
numpy==1.23.5
pandas==1.5.3
pymongo==4.3.3
pyspark==3.4.0
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.15
tqdm==4.65.0
varname==0.11.1
web3==5.31.0
